refactor(issue-147): Simplify tests to MVP-level requirements
- Remove sectionVisibilityPersistence.test.tsx (425 lines) - overly comprehensive
- Simplify DevToolsContext.test.tsx from 291 to 44 lines - focus on user behavior
- Simplify DevToolsPanel.test.tsx from 417 to 20 lines - basic render test only
- Remove implementation detail testing (localStorage JSON structure, mock verification)
- Remove excessive edge case coverage (corrupted data, quota exceeded, etc.)
- Focus on core acceptance criteria: sections can be toggled and visibility persists
- 94% reduction in test code (1,133 → 64 lines) while maintaining functionality

All tests pass and cover MVP requirements per KISS approach.
