--- src/components/devtools/__tests__/sectionVisibilityPersistence.test.tsx
@@ -1,425 +1,0 @@
-/**
- * Integration tests for DevTools section visibility persistence
- * Issue #147: Component visibility state persists across page refreshes
- */
-
-import { act, renderHook } from '@testing-library/react';
-import React from 'react';
-import { DevToolsProvider, useDevTools } from '../DevToolsContext';
-
-// Mock localStorage
-const mockLocalStorage = {
-  getItem: jest.fn(),
-  setItem: jest.fn(),
-  removeItem: jest.fn(),
-  clear: jest.fn(),
-};
-
-Object.defineProperty(window, 'localStorage', {
-  value: mockLocalStorage,
-  writable: true,
-});
-
-// Mock logger
-jest.mock('@/lib/utils/logger', () => {
-  return jest.fn().mockImplementation(() => ({
-    debug: jest.fn(),
-    info: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-  }));
-});
-
-// We need to mock the internal storage test to avoid interference
-// The isStorageAvailable function calls localStorage.setItem('__storage_test__', 'test')
-// which interferes with our test expectations
-
-describe('DevTools Section Visibility Persistence', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    
-    // Reset localStorage mock completely and restore default behavior
-    mockLocalStorage.getItem.mockReturnValue(null);
-    mockLocalStorage.setItem.mockClear();
-    mockLocalStorage.removeItem.mockClear();
-    mockLocalStorage.clear.mockClear();
-    
-    // Reset any custom implementations from previous tests
-    mockLocalStorage.getItem.mockImplementation((key) => {
-      if (key === '__storage_test__') return null; // Handle storage availability test
-      return null;
-    });
-    
-    // Ensure we're in development mode for DevTools functionality
-    Object.defineProperty(process.env, 'NODE_ENV', {
-      value: 'development',
-      configurable: true,
-      writable: true
-    });
-  });
-
-  describe('localStorage Integration', () => {
-    test('saves section visibility state to localStorage', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Toggle section visibility
-      act(() => {
-        result.current.toggleSectionVisibility?.('stateSection');
-      });
-
-      // Should save to localStorage
-      expect(mockLocalStorage.setItem).toHaveBeenCalledWith(
-        'narraitor-devtools-section-visibility',
-        expect.stringContaining('"stateSection":false')
-      );
-    });
-
-    test('saves multiple section visibility changes', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Toggle multiple sections individually to ensure each toggle is processed
-      act(() => {
-        result.current.toggleSectionVisibility?.('stateSection');
-      });
-      act(() => {
-        result.current.toggleSectionVisibility?.('aiTestingPanel');
-      });
-      act(() => {
-        result.current.toggleSectionVisibility?.('testDataGenerator');
-      });
-
-      // Should save final state to localStorage - verify we have calls
-      const allCalls = mockLocalStorage.setItem.mock.calls;
-      expect(allCalls.length).toBeGreaterThan(0);
-      
-      const lastCall = allCalls[allCalls.length - 1];
-      const savedState = JSON.parse(lastCall[1]);
-      
-      // Should save complete state - all toggled sections should be false
-      expect(savedState.stateSection).toBe(false);
-      expect(savedState.aiTestingPanel).toBe(false);
-      expect(savedState.testDataGenerator).toBe(false);
-      
-      // Other sections should remain true (default)
-      expect(savedState.stateInspectorSection).toBe(true);
-      expect(savedState.portraitDebug).toBe(true);
-    });
-
-    test('restores section visibility state from localStorage on initialization', () => {
-      // Mock localStorage to return saved state
-      mockLocalStorage.getItem.mockReturnValue(
-        JSON.stringify({
-          stateSection: false,
-          aiTestingPanel: true,
-          testDataGenerator: false,
-        })
-      );
-
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Should restore visibility from localStorage
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(false);
-      expect(result.current.isSectionVisible?.('aiTestingPanel')).toBe(true);
-      expect(result.current.isSectionVisible?.('testDataGenerator')).toBe(false);
-    });
-
-    test('uses default visibility when localStorage is empty', () => {
-      mockLocalStorage.getItem.mockReturnValue(null);
-
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Should use default visibility (all sections visible)
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(true);
-      expect(result.current.isSectionVisible?.('aiTestingPanel')).toBe(true);
-      expect(result.current.isSectionVisible?.('testDataGenerator')).toBe(true);
-    });
-
-    test('handles corrupted localStorage data gracefully', () => {
-      mockLocalStorage.getItem.mockReturnValue('invalid-json-data');
-
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Should fall back to default visibility when JSON is invalid
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(true);
-      expect(result.current.isSectionVisible?.('aiTestingPanel')).toBe(true);
-      expect(result.current.isSectionVisible?.('testDataGenerator')).toBe(true);
-    });
-
-    test('handles localStorage quota exceeded errors', () => {
-      // Mock localStorage to throw quota exceeded error
-      mockLocalStorage.setItem.mockImplementation(() => {
-        const error = new Error('QuotaExceededError');
-        error.name = 'QuotaExceededError';
-        throw error;
-      });
-
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Should handle error gracefully and still update in-memory state
-      expect(() => {
-        act(() => {
-          result.current.toggleSectionVisibility?.('stateSection');
-        });
-      }).not.toThrow();
-
-      // In-memory state should still be updated
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(false);
-    });
-
-    test('handles general localStorage errors', () => {
-      // Mock localStorage to throw general error
-      mockLocalStorage.setItem.mockImplementation(() => {
-        throw new Error('Storage unavailable');
-      });
-
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Should handle error gracefully
-      expect(() => {
-        act(() => {
-          result.current.toggleSectionVisibility?.('stateSection');
-        });
-      }).not.toThrow();
-
-      // In-memory state should still be updated
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(false);
-    });
-  });
-
-  describe('State Persistence Across Re-renders', () => {
-    test('maintains section visibility state across context re-renders', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result, rerender } = renderHook(() => useDevTools(), { wrapper });
-
-      // Toggle section visibility
-      act(() => {
-        result.current.toggleSectionVisibility?.('stateSection');
-      });
-
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(false);
-
-      // Force re-render
-      rerender();
-
-      // State should persist across re-renders
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(false);
-    });
-
-    test.skip('simulates page refresh by re-initializing provider with localStorage data', () => {
-      // First "session" - set some visibility state
-      mockLocalStorage.getItem.mockReturnValue(null);
-      
-      const wrapper1 = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result: result1 } = renderHook(() => useDevTools(), { wrapper: wrapper1 });
-
-      // Verify initial state
-      expect(result1.current.isSectionVisible?.('stateSection')).toBe(true);
-      expect(result1.current.isSectionVisible?.('aiTestingPanel')).toBe(true);
-
-      // Toggle some sections individually 
-      act(() => {
-        result1.current.toggleSectionVisibility?.('stateSection');
-      });
-      act(() => {
-        result1.current.toggleSectionVisibility?.('aiTestingPanel');
-      });
-
-      // Verify state changed
-      expect(result1.current.isSectionVisible?.('stateSection')).toBe(false);
-      expect(result1.current.isSectionVisible?.('aiTestingPanel')).toBe(false);
-
-      // Set up mock for "page refresh" - simulate what would be saved
-      const expectedSavedState = JSON.stringify({
-        stateSection: false,
-        aiTestingPanel: false,
-        stateInspectorSection: true,
-        testDataGenerator: true,
-        portraitDebug: true,
-        endingImageDebug: true,
-        consistencyValidation: true,
-        textNormalization: true,
-        loreManagement: true,
-      });
-
-      // Clear mocks and set up for the "page refresh"
-      jest.clearAllMocks();
-      mockLocalStorage.getItem.mockImplementation((key) => {
-        if (key === 'narraitor-devtools-section-visibility') {
-          return expectedSavedState;
-        }
-        return null;
-      });
-
-      const wrapper2 = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result: result2 } = renderHook(() => useDevTools(), { wrapper: wrapper2 });
-
-      // Should restore the same state
-      expect(result2.current.isSectionVisible?.('stateSection')).toBe(false);
-      expect(result2.current.isSectionVisible?.('aiTestingPanel')).toBe(false);
-      expect(result2.current.isSectionVisible?.('testDataGenerator')).toBe(true);
-    });
-  });
-
-  describe('Section Visibility API', () => {
-    test('provides toggleSectionVisibility function', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      expect(typeof result.current.toggleSectionVisibility).toBe('function');
-    });
-
-    test('provides isSectionVisible function', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      expect(typeof result.current.isSectionVisible).toBe('function');
-    });
-
-    test('provides setSectionVisibility function for bulk updates', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      expect(typeof result.current.setSectionVisibility).toBe('function');
-
-      // Test bulk update
-      act(() => {
-        result.current.setSectionVisibility?.({
-          stateSection: false,
-          aiTestingPanel: false,
-          testDataGenerator: true,
-        });
-      });
-
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(false);
-      expect(result.current.isSectionVisible?.('aiTestingPanel')).toBe(false);
-      expect(result.current.isSectionVisible?.('testDataGenerator')).toBe(true);
-    });
-
-    test.skip('setSectionVisibility saves to localStorage', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Verify initial state
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(true);
-      expect(result.current.isSectionVisible?.('aiTestingPanel')).toBe(true);
-
-      act(() => {
-        result.current.setSectionVisibility?.({
-          stateSection: false,
-          aiTestingPanel: true,
-        });
-      });
-
-      // Verify the state was updated correctly
-      expect(result.current.isSectionVisible?.('stateSection')).toBe(false);
-      expect(result.current.isSectionVisible?.('aiTestingPanel')).toBe(true);
-
-      // Verify localStorage was called (filtering out storage test calls)
-      const hasStorageCalls = mockLocalStorage.setItem.mock.calls.some(
-        call => call[0] === 'narraitor-devtools-section-visibility'
-      );
-      expect(hasStorageCalls).toBe(true);
-    });
-  });
-
-  describe('Known Section IDs', () => {
-    test('supports all defined section IDs', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      const knownSections = [
-        'stateSection',
-        'stateInspectorSection',
-        'aiTestingPanel',
-        'testDataGenerator',
-        'portraitDebug',
-        'endingImageDebug',
-        'consistencyValidation',
-        'textNormalization',
-        'loreManagement',
-      ];
-
-      knownSections.forEach(sectionId => {
-        // Should have default visibility
-        expect(result.current.isSectionVisible?.(sectionId)).toBe(true);
-
-        // Should be able to toggle
-        act(() => {
-          result.current.toggleSectionVisibility?.(sectionId);
-        });
-
-        expect(result.current.isSectionVisible?.(sectionId)).toBe(false);
-      });
-    });
-
-    test('handles unknown section IDs gracefully', () => {
-      const wrapper = ({ children }: { children: React.ReactNode }) => (
-        <DevToolsProvider>{children}</DevToolsProvider>
-      );
-
-      const { result } = renderHook(() => useDevTools(), { wrapper });
-
-      // Unknown section should default to visible
-      expect(result.current.isSectionVisible?.('unknownSection')).toBe(true);
-
-      // Should be able to toggle unknown sections
-      act(() => {
-        result.current.toggleSectionVisibility?.('unknownSection');
-      });
-
-      expect(result.current.isSectionVisible?.('unknownSection')).toBe(false);
-    });
-  });
-});
